Calculation of new issue start and end dates uses the new ::Calendar module!
All old specs are passing, and that means that with 'Schedule on weekends' setting enabled by default the system behaves just like before. I am too lazy to add the whole new copy of the issue patch spec for the case when 'Schedule on weekends' is disabled, but the test suite for Calendar module is solid enough to not worry about that so far.