Removed issue attribute 'estimated duration' and setting 'calculate duration'.
Issue already has 'duration', that is computed as difference between due and start dates. The plugin calculates it in working days. Because of that, there is no need to store 'estimated duration' attribute. Also, changing 'Work on weekends' setting on plugin would require recalculation of estimated durations of all issues, which I rather would not do.